Update M2 Manage manufacturing processes.md
Before adding this step for the Route relation field, when I clicked Save I got an error message stating that Route Relation must be specified.  Therefore, I chose the route (eBook route) previously created and was able to click Save.  The Route code field was already displayed and was set to Route and not All.